pyproject.toml
[project]
name = "demo-surin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipython>=9.6.0",
    "langchain>=0.3.27",
    "langchain-community==0.3.17",
    "langchain-core>=0.3.76",
    "langchain-tavily>=0.2.11",
    "langchain-upstage>=0.7.3",
    "langgraph>=0.6.8",
    "langgraph-cli[inmem]>=0.4.2",
    "numpy==1.26.4",
    "openai>=1.109.1",
    "psycopg2-binary>=2.9.10",
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "tavily-python>=0.7.12",
    "tqdm>=4.67.1",
    "torch==2.2.2",
    "qdrant-client>=1.12.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
